% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglda.R
\docType{package}
\name{biglda-package}
\alias{biglda-package}
\alias{biglda}
\title{biglda-package}
\description{
Package for fast LDA topic modelling for big corpora.
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
