% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{density}
\alias{density}
\title{Fast Implementation of Cao et al. 2009}
\usage{
density(model)
}
\arguments{
\item{model}{A LDA model (S4 class from topicmodels package).}
}
\description{
The function picks up the suggestion of Cao et al. 2009 for a density-based 
method for LDA model selection and offers an implementation that is 
substantially faster than ldatuning.
}
