% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileops.R
\name{save_document_topics}
\alias{save_document_topics}
\alias{load_document_topics}
\title{Process large Document-Topics-Matrices}
\usage{
save_document_topics(model, destfile = tempfile())

load_document_topics(filename, verbose = TRUE)
}
\arguments{
\item{model}{A rJava object `ParallelTopicModel` or inheriting from it.}

\item{destfile}{Path to a file for temporarily saving data on documents and
topics.}

\item{filename}{Path to file with data exported from Java.}

\item{verbose}{A `logical` value, whether to show progress messages.}
}
\description{
Process large Document-Topics-Matrices
}
\examples{
m <- system.file(package = "biglda", "extdata", "mallet", "lda_mallet.bin") |>
  mallet_load_topicmodel()
fname <- save_document_topics(m)
y <- load_document_topics(fname)
dim(y)
}
