% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensim.R
\name{dtm_as_bow}
\alias{dtm_as_bow}
\title{Convert DocumentTermMatrix to Gensim corpus.}
\usage{
dtm_as_bow(dtm)
}
\arguments{
\item{dtm}{A Document-Term-Matrix.}
}
\description{
Convert DocumentTermMatrix to Gensim corpus.
}
\details{
The input to gensim's LDA modelling methods is a representation of
  corpora in a data format denoted as "BOW". This utility function
  `dtm_as_bow()` turns a sparse matrix (class `simple_triplet_matrix`) into
  the bow input format required by gensim.
}
\examples{
if (requireNamespace("reticulate") && reticulate::py_module_available("gensim")){
  library(polmineR)
  gensim <- reticulate::import("gensim")
  use("RcppCWB", corpus = "REUTERS")

  bow <- corpus("REUTERS") \%>\%
    split(s_attribute = "id") \%>\%
    as.DocumentTermMatrix(p_attribute = "word", verbose = FALSE) \%>\%
    dtm_as_bow()
}   
}
\author{
Andreas Blaette
}
