% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{FastArun2010}
\alias{FastArun2010}
\alias{FastArun2010,TopicModel-method}
\alias{FastArun2010,matrix-method}
\alias{FastArun2010,jobjRef-method}
\title{Fast Implementation of Arun 2014}
\usage{
FastArun2010(x, ...)

\S4method{FastArun2010}{TopicModel}(x, doclengths)

\S4method{FastArun2010}{matrix}(x, gamma, doclengths)

\S4method{FastArun2010}{jobjRef}(x)
}
\arguments{
\item{x}{A topic model.}

\item{...}{Further arguments}

\item{doclengths}{Integer vector.}

\item{gamma}{The gamma matrix.}
}
\description{
Fast Implementation of Arun 2014
}
\examples{
\dontrun{
if (!mallet_is_installed()) mallet_install()
library(polmineR)
speeches <- polmineR::as.speeches("GERMAPARLMINI", s_attribute_name = "speaker")
instance_list <- as.instance_list(speeches)
lda <- BigTopicModel(n_topics = 25L, alpha_sum = 5.1, beta = 0.1)
lda$addInstances(instance_list)
lda$setNumThreads(1L)
lda$setNumIterations(150L)
lda$estimate()
lda2 <- as_LDA(lda)
FastArun2010(lda2, doclengths = summary(speeches)[["size"]])
}
if (!mallet_is_installed()) mallet_install()
fname <- system.file(package = "biglda", "extdata", "mallet", "lda_mallet2.bin")
lda <- mallet_load_topicmodel(fname)
FastArun2010(lda)
}
