% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{FastCao2009}
\alias{FastCao2009}
\alias{FastCao2009,TopicModel-method}
\alias{FastCao2009,matrix-method}
\alias{FastCao2009,jobjRef-method}
\title{Fast Implementation of Cao et al. 2009}
\usage{
FastCao2009(x)

\S4method{FastCao2009}{TopicModel}(x)

\S4method{FastCao2009}{matrix}(x)

\S4method{FastCao2009}{jobjRef}(x)
}
\arguments{
\item{x}{A LDA model (S4 class from topicmodels package).}
}
\description{
The function picks up the suggestion of Cao et al. 2009 for a density-based 
method for LDA model selection and offers an implementation that is 
substantially faster than ldatuning.
}
\examples{
if (!mallet_is_installed()) mallet_install()
fname <- system.file(package = "biglda", "extdata", "mallet", "lda_mallet2.bin")
lda <- mallet_load_topicmodel(fname)
lda2 <- as_LDA(lda)
FastCao2009(lda2)
}
