% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Get topic model diagnostics.}
\usage{
get_diagnostics(x, n = 50L, verbose = TRUE)
}
\arguments{
\item{x}{An instance of the RTopicModel class (java object).}

\item{n}{An `integer` value, the number of the top words that will be evaluated.}

\item{verbose}{A `logical` value, whether to show progress messages.}
}
\value{
A `data.table`.
}
\description{
The MALLET topic model toolkit includes a class `TopicModelDiagnostics` able
to prepare a set of metrics on the topics of a topic model. The function
`get_diagnostics()` will return a `data.table` with these
diagnostics. See the [Mallet
documentation](http://mallet.cs.umass.edu/diagnostics.php) for an explanation
of the metrics.
}
\examples{
f <- system.file(package = "biglda", "extdata", "mallet", "lda_mallet.bin")
model <- mallet_load_topicmodel(f)
diagnostics <- get_diagnostics(model)
head(diagnostics)
}
