% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_formatted_filesize}
\alias{get_formatted_filesize}
\title{Get size of file and return formatted value}
\usage{
get_formatted_filesize(x)
}
\arguments{
\item{x}{Path to a file}
}
\description{
Get size of file and return formatted value
}
