% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mallet_dir}
\alias{get_mallet_dir}
\title{Get directory with Mallet jar files.}
\usage{
get_mallet_dir()
}
\description{
This utility function checks two options where Mallet jars might be: (a) The
directory indicated by the environment variable 'MALLET_DIR' and the
directory 'extdata/java' within the installed biglda package.
}
