% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_LDA.R
\docType{class}
\name{LDA_Mallet-class}
\alias{LDA_Mallet-class}
\alias{LDA_Gensim-class}
\title{LDA Mallet Class.}
\description{
LDA Mallet Class.
}
\section{Slots}{

\describe{
\item{\code{doclengths}}{An `integer` vector with document lengths.}
}}

