% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet.R
\docType{class}
\name{LDA_matched-class}
\alias{LDA_matched-class}
\alias{mallet_make_instance_list}
\alias{as.instance_list}
\alias{as_LDA}
\alias{mallet_instance_list_store}
\alias{mallet_instance_list_load}
\alias{mallet_load_topicmodel}
\alias{mallet_is_installed}
\alias{mallet_get_version}
\alias{mallet_install}
\title{Interface to mallet topicmodelling.}
\usage{
mallet_make_instance_list(
  x,
  p_attribute = "word",
  phrases = NULL,
  terms_to_drop = tm::stopwords("de"),
  mc = TRUE,
  verbose = TRUE
)

as.instance_list(x, p_attr = "word", verbose = TRUE)

as_LDA(x, verbose = TRUE, beta = NULL, gamma = NULL)

mallet_instance_list_store(x, filename = tempfile())

mallet_instance_list_load(filename)

mallet_load_topicmodel(filename)

mallet_is_installed()

mallet_get_version()

mallet_install(url = "http://mallet.cs.umass.edu/dist/mallet-2.0.8.tar.gz")
}
\arguments{
\item{x}{A \code{partition_bundle} object.}

\item{p_attribute}{The p_attribute to use, typically "word" or "lemma".}

\item{phrases}{A \code{phrases} object (S4 class from polmineR) that will be
used to concatenate phrases.}

\item{terms_to_drop}{stopwords}

\item{mc}{A \code{logical} value, whether to use multicore.}

\item{verbose}{A \code{logical} value, whether to be verbose.}

\item{p_attr}{a p-attribute}

\item{beta}{The beta matrix for a topic model.}

\item{gamma}{The gamma matrix for a topic model.}

\item{filename}{Where to store the Java-object.}

\item{url}{url of mallet}

\item{...}{further parameters}
}
\description{
Functionality to support the following workflow (see examples): (a) Turn
\code{partition_bundle}-object into mallet instance list, (b) store the
resulting \code{jobjRef}-object, (c) run mallet topic modelling and (d)
turn ParallelTopicModel Java object into \code{LDA_Gibbs} object from
package \code{topicmodels}.
}
\details{
The \code{as_LDA()}-function will turn an estimated topic model
  prepared using 'mallet' into a \code{LDA_Gibbs} object as defined in the
  \code{topicmodels} package. This may be useful for using topic model
  evaluation tools available for the \code{LDA_Gibbs} class, but not for the
  immediate output of malled topicmodelling. Note that the gamma matrix is
  normalized and smoothed, the beta matrix is the logarithmized matrix of
  normalized and smoothed values obtained from the input mallet topic model.

The function \code{mallet_instance_list_load} will load a Java
  InstanceList object that has been saved to disk (e.g. by using the
  \code{mallet_instance_list_store} function). The return value is a
  \code{jobjRef} object. Internally, the function reuses code of the function
  \code{load.mallet.instances} from the R package \code{mallet}.

The function \code{mallet_load_topicmodel} will load a topic model
  created using mallet into memory.

The \code{mallet_is_installed} utility will return a logical value
  whether mallet is installed or not.
}
\examples{
 
# Preparations: Create instance list

if (!mallet_is_installed()) mallet_install()

polmineR::use("polmineR")
speeches <- polmineR::as.speeches("GERMAPARLMINI", s_attribute_name = "speaker")

  instance_list <- as.instance_list(speeches)

  lda <- rJava::.jnew("cc/mallet/topics/ParallelTopicModel", 25L, 5.1, 0.1)
  lda$addInstances(instance_list)
  lda$setNumThreads(1L)
  lda$setTopicDisplay(50L, 10L)
  lda$setNumIterations(150L)
  lda$estimate()
  # lda$write(rJava::.jnew("java/io/File", destfile))

# Load topicmodel and turn it into LDA_Gibbs

# mallet_lda <- mallet_load_topicmodel(destfile)
topicmodels_lda <- as_LDA(lda)

library(polmineR)
use("polmineR")
speeches <- as.speeches("GERMAPARLMINI", s_attribute_name = "speaker")
speeches_instance_list <- as.instance_list(speeches, p_attr = "word")
}
