% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.instance_list.R
\name{mallet_get_topic_model_diagnostics}
\alias{mallet_get_topic_model_diagnostics}
\title{Get topic model diagnostics.}
\usage{
mallet_get_topic_model_diagnostics(x, n = 100L)
}
\arguments{
\item{x}{An instance of the RTopicModel class (java object).}

\item{n}{Number of the top words that will be evaluated.}
}
\description{
The MALLET topic model toolkit includes a class \code{TopicModelDiagnostics}
able to prepare a set of metrics on the topics of a topic model. The function
\code{mallet_get_topic_model_diagnostics} will return a \code{data.table}
with these diagnostics. See the
\href{http://mallet.cs.umass.edu/diagnostics.php}{mallet documentation} for an
explanation of the metrics.
}
\examples{
\dontrun{
dt <- mallet_get_topic_model_diagnostics(lda)
}
}
