% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mallet_install}
\alias{mallet_install}
\alias{mallet_get_version}
\alias{mallet_is_installed}
\title{Functions to manage mallet installation}
\usage{
mallet_install(url = "http://mallet.cs.umass.edu/dist/mallet-2.0.8.tar.gz")

mallet_get_version()

mallet_is_installed()
}
\arguments{
\item{url}{url of mallet}
}
\description{
Functions to manage mallet installation
}
\details{
The \code{mallet_install} function will downlpad the mallet Java package
  and put it into the java directory of the biglda R package.

The \code{mallet_get_version} function will return a
  \code{numeric_version} object of mallet is installed, or \code{NULL} if
  mallet is not available.

The \code{mallet_is_installed} utility will return a logical value
  whether mallet is installed or not.
}
\examples{
if (isFALSE(mallet_is_installed())) mallet_install()
mallet_get_version()

}
