% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,TopicModel-method}
\alias{plot.metrics}
\title{Evaluate topic model.}
\usage{
metrics(x, ...)

\S4method{metrics}{TopicModel}(x, verbose = TRUE)

\method{plot}{metrics}(x, ...)
}
\arguments{
\item{x}{Input object of class inheriting from `TopicModel`.}

\item{...}{Further arguments (unused).}

\item{verbose}{A `logical` value, whether to output progress messages.}
}
\description{
Basic plot to visualize metrics. For a nicer, more elaborate plot, see
`ldatuning::FindTopicsNumber_plot()`
}
\examples{
lda <- system.file(package = "biglda", "extdata", "mallet", "lda_mallet.bin") |>
  mallet_load_topicmodel() |>
  as_LDA()
m <- metrics(lda)
if (interactive()) plot(m)
}
