% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_token_stream}
\alias{write_token_stream}
\alias{write_token_stream2}
\alias{BigCao2009}
\alias{BigArun2010}
\alias{BigDeveaud2014}
\title{Rcpp/RcppCWB implementation for writing token stream}
\usage{
write_token_stream(
  corpus,
  p_attribute,
  s_attribute,
  registry,
  attribute_type,
  filename
)

write_token_stream2(
  corpus,
  p_attribute,
  s_attribute,
  strucs,
  registry,
  attribute_type,
  filename
)

BigCao2009(X)

BigArun2010(beta, gamma, doclengths)

BigDeveaud2014(beta)
}
\arguments{
\item{corpus}{corpus.}

\item{p_attribute}{X}

\item{s_attribute}{X}

\item{registry}{X}

\item{attribute_type}{X}

\item{filename}{X}

\item{X}{The beta matrix of an LDA topic model.}

\item{beta}{The beta matrix.}

\item{gamma}{The gamma matrix.}

\item{doclengths}{A `vector` with document lengths (number of tokens).}
}
\description{
Rcpp/RcppCWB implementation for writing token stream

Rcpp/RcppCWB implementation for writing token stream

RcppArmadillo implementation of topic model metrics
}
\references{
\url{https://dn.revuesonline.com/gratuit/DN17_1_06_Deveau.pdf}
}
