% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{J}
\alias{.jnew}
\alias{.jcast}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rJava}{\code{\link[rJava]{.jcast}}, \code{\link[rJava]{.jnew}}, \code{\link[rJava]{J}}}
}}

