% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R, R/gamma.R
\docType{methods}
\name{B}
\alias{B}
\alias{B,TopicModel-method}
\alias{B<-}
\alias{B<-,TopicModel-method}
\alias{G}
\alias{G,TopicModel-method}
\alias{G<-}
\alias{G<-,TopicModel-method}
\title{Access and modify slot for beta matrix.}
\usage{
B(x, ...)

\S4method{B}{TopicModel}(x)

B(x) <- value

\S4method{B}{TopicModel}(x) <- value

G(x, ...)

\S4method{G}{TopicModel}(x)

G(x) <- value

\S4method{G}{TopicModel}(x) <- value
}
\arguments{
\item{x}{An object inheriting from the `TopicModel` class.}

\item{...}{Further arguments.}

\item{value}{A matrix to assign (beta or gamma).}
}
\description{
Access and modify slot for beta matrix.
}
