% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_mallet_input}
\alias{write_mallet_input}
\title{biglda::write_mallet_input(
  corpus = "ALBB",
  s_attribute = "article_id",
  p_attribute = "word",
  strucs = 0L:(size - 1L),
  registry = corpus_registry_dir("ALBB"),
  filename = fname
)}
\usage{
write_mallet_input(
  corpus,
  p_attribute,
  s_attribute,
  strucs,
  registry,
  filename
)
}
\description{
biglda::write_mallet_input(
  corpus = "ALBB",
  s_attribute = "article_id",
  p_attribute = "word",
  strucs = 0L:(size - 1L),
  registry = corpus_registry_dir("ALBB"),
  filename = fname
)
}
