% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seededlda.R
\name{write_topic_priors}
\alias{write_topic_priors}
\title{Write topic priors file}
\usage{
write_topic_priors(x, file)
}
\description{
Write topic priors file
}
\examples{
topic_priors <- list(
  "b" = c("oil", "barrel"),
  "2" = c("Bahrain")
)
fname <- tempfile(fileext = ".txt")
write_topic_priors(x = topic_priors, file = fname)
}
