% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{FastDeveaud2014}
\alias{FastDeveaud2014}
\title{Fast Implementation of Deveaud 2014}
\usage{
FastDeveaud2014(model, cl = parallel::detectCores() - 1L)
}
\arguments{
\item{model}{A LDA model.}

\item{cl}{Number of cores to use}
}
\description{
Fast Implementation of Deveaud 2014
}
