% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensim.R
\name{gensim_ldamodel_load}
\alias{gensim_ldamodel_load}
\title{Load gensim topic model.}
\usage{
gensim_ldamodel_load(modeldir, modelname)
}
\arguments{
\item{modeldir}{Directory where a gensim LDA topic model has been saved.}

\item{modelname}{Name of a gensim LDA topic model. The data for a model}
}
\description{
Load gensim topic model.
}
\details{
Use `gensim_ldamodel_load()` to load an ldamodel computed by gensim.
  The return value is a `LdaModel` Python object that can serve as input to
  functions or that can be processed using the `reticulate` package.
}
\author{
Andreas Blaette
}
