% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mallet_install}
\alias{mallet_install}
\alias{mallet_get_version}
\alias{mallet_is_installed}
\title{Functions to manage mallet installation}
\usage{
mallet_install(
  url =
    "https://github.com/mimno/Mallet/releases/download/v202108/Mallet-202108-bin.tar.gz",
  verbose = TRUE
)

mallet_get_version()

mallet_is_installed()
}
\arguments{
\item{url}{url of mallet}

\item{verbose}{A `logical` value, whether to output progress messages.}
}
\description{
Functions to manage mallet installation
}
\details{
`mallet_install()` will download the mallet Java package
  and put it into the java directory of the biglda R package.

`mallet_get_version()` will return a `numeric_version` object of
  mallet is installed, or `NULL` if mallet is not available.

`mallet_is_installed()` utility will return a `logical` value
  whether mallet is installed or not.
}
\examples{
if (isFALSE(mallet_is_installed())) mallet_install()
mallet_get_version()

}
