% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{as_LDA}
\alias{as_LDA}
\title{Convert mallet LDA to topicanalysis class}
\usage{
as_LDA(x, verbose = TRUE, beta = NULL, gamma = NULL)
}
\arguments{
\item{x}{A mallet topic model (ParallelTopicModel).}

\item{verbose}{A \code{logical} value, whether to output progress messages.}

\item{beta}{The beta matrix for a topic model.}

\item{gamma}{The gamma matrix for a topic model.}
}
\description{
Convert mallet LDA to topicanalysis class
}
\details{
The \code{as_LDA()}-function will turn an estimated topic model
  prepared using 'mallet' into a \code{LDA_Gibbs} object as defined in the
  \code{topicmodels} package. This may be useful for using topic model
  evaluation tools available for the \code{LDA_Gibbs} class, but not for the
  immediate output of malled topicmodelling. Note that the gamma matrix is
  normalized and smoothed, the beta matrix is the logarithmized matrix of
  normalized and smoothed values obtained from the input mallet topic model.
}
