% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{as_LDA}
\alias{as_LDA}
\title{Convert mallet LDA to topicanalysis class}
\usage{
as_LDA(x, verbose = TRUE, beta = NULL, gamma = NULL)
}
\arguments{
\item{x}{A mallet topic model (ParallelTopicModel).}

\item{verbose}{A `logical` value, whether to output progress messages.}

\item{beta}{The beta matrix for a topic model.}

\item{gamma}{The gamma matrix for a topic model.}
}
\description{
Convert mallet LDA to topicanalysis class
}
\details{
The `as_LDA()`-function will turn an estimated topic model prepared
  using 'mallet' into a `LDA_Mallet` object that inherits from classes
  defined in the `topicmodels` package. This may be useful for using topic
  model evaluation tools available for the `LDA_Gibbs` class, but not for the
  immediate output of malled topicmodelling. Note that the gamma matrix is
  normalized and smoothed, the beta matrix is the logarithmized matrix of
  normalized and smoothed values obtained from the input mallet topic model.
}
\examples{
data_dir <- system.file(package = "biglda", "extdata", "mallet")
BTM <- mallet_load_topicmodel(
  instancefile = file.path(data_dir, "instance_list.mallet"),
  statefile = file.path(data_dir, "lda_mallet.gz")
)

LDA <- as_LDA(BTM)

# Avoid memory limitations as follows
LDA2 <- as_LDA(BTM, beta = matrix(), gamma = matrix())

B(LDA2) <- save_word_weights(BTM, minimized = TRUE) |>
  load_word_weights(minimized = TRUE) |>
  exp()
  
G(LDA2) <- save_document_topics(BTM) |>
  load_document_topics()
}
