% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileops.R
\name{mallet_load_topicmodel}
\alias{mallet_load_topicmodel}
\alias{ParallelTopicModel}
\title{Instantiate and load mallet topicmodel}
\usage{
mallet_load_topicmodel(filename)

ParallelTopicModel(n_topics = 25L, alpha_sum = 5.1, beta = 0.1)
}
\arguments{
\item{filename}{Filename of a mallet topic model (ParallelTopicModel).}

\item{n_topics}{Number of topics (\code{integer} value).}

\item{alpha_sum}{Passed into constructor.}

\item{beta}{Passet into constructor.}
}
\description{
Instantiate and load mallet topicmodel
}
\details{
The function \code{mallet_load_topicmodel} will load a topic model
  created using mallet into memory.

The \code{ParallelTopicModel} function will instantial a Java class
  object with the same name from the mallet package, see the
  \href{http://mallet.cs.umass.edu/api/cc/mallet/topics/ParallelTopicModel.html}{mallet
  documentation} of the class.
}
\examples{
pta <- ParallelTopicModel()
destfile <- tempfile()
pta$write(rJava::.jnew("java/io/File", destfile))
pta_reloaded <- mallet_load_topicmodel(destfile)
}
