% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BigCao2009}
\alias{BigCao2009}
\alias{BigArun2010}
\alias{BigDeveaud2014}
\title{RcppArmadillo implementation of topic model metrics}
\usage{
BigCao2009(X)

BigArun2010(beta, gamma, doclengths)

BigDeveaud2014(beta)
}
\arguments{
\item{X}{The beta matrix of an LDA topic model.}

\item{beta}{The beta matrix.}

\item{gamma}{The gamma matrix.}

\item{doclengths}{A `vector` with document lengths (number of tokens).}
}
\description{
RcppArmadillo implementation of topic model metrics
}
\references{
\url{https://dn.revuesonline.com/gratuit/DN17_1_06_Deveau.pdf}
}
