% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmlight.R
\name{svmlight}
\alias{svmlight}
\alias{svmlight,DocumentTermMatrix-method}
\title{Export Matrix as svmlight format}
\usage{
svmlight(x, ...)

\S4method{svmlight}{DocumentTermMatrix}(x, filename, verbose = TRUE)
}
\arguments{
\item{x}{Input object, `DocumentTermMatrix` for the time being.}

\item{...}{Additional arguments.}

\item{filename}{Path to a file.}

\item{verbose}{A `logical` value, whether to output progress messages.}
}
\description{
Export Matrix as svmlight format
}
\examples{
library(polmineR)
use("RcppCWB")

svmlight_file <- tempfile()
corpus("REUTERS") \%>\%
  as.DocumentTermMatrix(p_attribute = "word", s_attribute = "id") \%>\%
  svmlight(filename = svmlight_file)
}
